--- MarkEditMac/Sources/AppExtension/AppExtensionConfig.swift
@@ -1,124 +1,0 @@
-//
-//  AppExtensionConfig.swift
-//  MarkEditMac
-//
-//  Created by cyan on 7/11/26.
-//
-
-import Foundation
-import MarkEditCore
-import MarkEditKit
-
-/// App-side state for editor extensions, stored as "extensions.json".
-///
-/// Mirrors `AppRuntimeConfig`, the remote catalog is fetched by `AppExtensionRegistry`.
-enum AppExtensionConfig {
-  struct Definition: Codable {
-    let __schema: String?
-    let registryURL: String?
-    let updateCheck: UpdateCheck?
-    let updateStrategy: UpdateStrategy?
-    let installed: [Installed]?
-
-    enum CodingKeys: String, CodingKey {
-      case __schema = "$schema"
-      case registryURL = "registry.url"
-      case updateCheck = "registry.updateCheck"
-      case updateStrategy = "registry.updateStrategy"
-      case installed
-    }
-  }
-
-  /// How often to check for updates.
-  enum UpdateCheck: String, Codable {
-    case never
-    case onLaunch
-    case daily
-    case weekly
-  }
-
-  /// What to do once an update is found.
-  enum UpdateStrategy: String, Codable {
-    case manual
-    case prompt
-    case automatic
-  }
-
-  /// A single installed extension, array order is injection order.
-  struct Installed: Codable, Equatable {
-    let id: String
-    let version: String?
-    let url: String?
-    let sha256: String?
-    let file: String
-    let enabled: Bool
-    let updateCheck: UpdateCheck?
-    let installDate: String?
-  }
-
-  /// Registry index url, overridable to a mirror.
-  static var registryURL: URL? {
-    let string = currentDefinition?.registryURL ?? Constants.defaultRegistryURL
-    return URL(string: string)
-  }
-
-  static var updateCheck: UpdateCheck {
-    currentDefinition?.updateCheck ?? .weekly
-  }
-
-  static var updateStrategy: UpdateStrategy {
-    // Never automatic by default, injected code is arbitrary JavaScript
-    currentDefinition?.updateStrategy ?? .prompt
-  }
-
-  static var installed: [Installed] {
-    currentDefinition?.installed ?? []
-  }
-
-  static var defaultContents: String {
-    encode(definition: defaultDefinition)?.toString() ?? ""
-  }
-}
-
-// MARK: - Private
-
-private extension AppExtensionConfig {
-  enum Constants {
-    static let defaultRegistryURL = "https://raw.githubusercontent.com/MarkEdit-app/extensions/main/index.json"
-    static let schemaURL = "https://raw.githubusercontent.com/MarkEdit-app/schemas/main/extensions.json"
-  }
-
-  static let fileData = try? Data(contentsOf: AppCustomization.extensions.fileURL)
-
-  static let defaultDefinition = Definition(
-    __schema: Constants.schemaURL,
-    registryURL: Constants.defaultRegistryURL,
-    updateCheck: .weekly,
-    updateStrategy: .prompt,
-    installed: []
-  )
-
-  static let currentDefinition: Definition? = {
-    guard let fileData else {
-      Logger.log(.error, "Missing extensions.json to proceed")
-      return nil
-    }
-
-    guard let definition = try? JSONDecoder().decode(Definition.self, from: fileData) else {
-      Logger.log(.error, "Invalid json object was found: \(fileData)")
-      return nil
-    }
-
-    return definition
-  }()
-
-  static func encode(definition: Definition) -> Data? {
-    let encoder = JSONEncoder()
-    encoder.outputFormatting = [.prettyPrinted, .sortedKeys, .withoutEscapingSlashes]
-
-    let jsonData = try? encoder.encode(definition)
-    Logger.assert(jsonData != nil, "Failed to encode object: \(definition)")
-
-    return jsonData
-  }
-}
